pyproject.toml
[tool.poetry]
name = "relayer"
version = "0.1.0"
description = ""
authors = ["Arnaud Sene <arnaud.sene@pm.me>"]
readme = "README.md"
packages = [{include = "relayer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
web3 = "^6.19.0"
python-dotenv = "^1.0.1"
pytest-mock = "^3.14.0"
attributedict = "^0.3.0"
tomli = "^2.0.1"
jinja2 = "^3.1.4"
tqdm = "^4.66.4"
aio-pika = "^9.4.3"
pydantic = "^2.9.2"
plyvel-ci = "^1.5.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
ipython = "^8.24.0"
pytest-asyncio = "^0.23.7"
web3 = {extras = ["tester"], version = "^6.19.0"}


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
tox = "^4.22.0"
mypy = "^1.12.0"
flake8 = "^7.1.1"
pydocstyle = "^6.3.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
